+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Master (Unreleased)
 
+
+
 ### Breaking Changes
 
 #### Changes for THING register
@@ -20,6 +22,7 @@ register new thing with Emacs things, functions, syntax descriptions or regexp p
 * [#110](https://github.com/meow-edit/meow/issues/110) Only disable hint overlay for modes in `meow-expand-exclude-mode-list`.
 * Add custom variable `meow-motion-remap-prefix.`
 * Remove `dash.el` and `s.el` from dependencies.
+* Add more defaults to `meow-mode-state-list`.
 
 ## 1.1.1 (2021-12-06)