+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Master (Unreleased)
 
-
+## 1.2.0 (2021-12-16)
 
 ### Breaking Changes